move username generator to utils
