+++ modules/util/RandomUtil.js
@@ -1,13 +1,51 @@
+/**
+ * @const
+ */
+var ALPHANUM = '0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ';
+
+/**
+ * Generates random int within the range [min, max]
+ * @param min the minimum value for the generated number
+ * @param max the maximum value for the generated number
+ * @returns random int number
+ */
+function randomInt(min, max) {
+    return Math.floor(Math.random() * (max - min + 1)) + min;
+}
+
+/**
+ * Get random element from array or string.
+ * @param {Array|string} arr source
+ * @returns array element or string character
+ */
+function randomElement(arr) {
+    return arr[randomInt(0, arr.length -1)];
+}
+
+/**
+ * Generate random alphanumeric string.
+ * @param {number} length expected string length
+ * @returns {string} random string of specified length
+ */
+function randomAlphanumStr(length) {
+    var result = '';
+
+    for (var i = 0; i < length; i += 1) {
+        result += randomElement(ALPHANUM);
+    }
+
+    return result;
+}
 
 /**
  * Generates random hex number within the range [min, max]
- * @param max the maximum value for the generated number
  * @param min the minimum value for the generated number
+ * @param max the maximum value for the generated number
  * @returns random hex number
  */
 function rangeRandomHex(min, max)
 {
-    return Math.floor(Math.random() * (max - min) + min).toString(16);
+    return randomInt(min, max).toString(16);
 }
 
 /**
@@ -31,7 +69,11 @@ var RandomUtil = {
      */
     random12digitsHex: function() {
         return rangeRandomHex(17592186044416, 281474976710655);
-    }
+    },
+
+    randomElement: randomElement,
+
+    randomAlphanumStr: randomAlphanumStr
 };
 
 module.exports = RandomUtil;
